switch to 64-bit types

+++ QSSpotlightPlugIn_Action.m
@@ -58,7 +58,7 @@
 	[results removeObject:searching];
 	NSString *resultPath = nil;
 	// fast enumeration is not recommended for NSMetadataQuery
-	for (int i = 0; i < [query resultCount]; i++) {
+	for (NSUInteger i = 0; i < [query resultCount]; i++) {
 		// get the path and create a QSObject with it
 		resultPath = [[query resultAtIndex:i] valueForAttribute:NSMetadataItemPathKey];
 		[results addObject:[QSObject fileObjectWithPath:resultPath]];
